login check implemented

+++ application/controllers/moderate.php
@@ -27,21 +27,33 @@ class moderate extends Controller {
 
 	public function verifydetails($id,$moderationid){
 
+		if(isset($_SESSION['login']) && ($_SESSION['email'] == ADMIN_EMAIL)){
 			$data = $this->model->getModifedDetails($id,$moderationid);
 			$data["original"] = $this->model->getJsonFor($id);
 			($data) ? $this->view('moderate/showDifference', $data) : $this->view('error/index');
+		}
+		else
+		{
+			$this->redirect('user/login');
+		}
 	}
 
 	public function discard($id,$moderationid){
 
-		$dbh = $this->model->db->connect(DB_NAME);
-		$data = array();
-		$data["id"] = $id;
-		$this->model->db->deleteDataFromModeration($id,$moderationid,$dbh);
-		$message = $this->model->bindVariablesToString(MDR_DISCARD_MSG,$data);
-		//do we need to send an email to a person who modified the details?
-		//decision has to be taken
-		($data) ? $this->view('moderate/discard', $message) : $this->view('error/index');
+		if(isset($_SESSION['login']) && ($_SESSION['email'] == ADMIN_EMAIL)){
+			$dbh = $this->model->db->connect(DB_NAME);
+			$data = array();
+			$data["id"] = $id;
+			$this->model->db->deleteDataFromModeration($id,$moderationid,$dbh);
+			$message = $this->model->bindVariablesToString(MDR_DISCARD_MSG,$data);
+			//do we need to send an email to a person who modified the details?
+			//decision has to be taken
+			($data) ? $this->view('moderate/discard', $message) : $this->view('error/index');
+		}
+		else
+		{
+			$this->redirect('user/login');
+		}		
 	}
 
 }
